Navigate to filled cell using number bar
If the selected cell is already filled in with a correct value,
and you click on a different number's button,
and there is a cell that is already filled with that different number,
the selected cell will switch to one of the cells with that different
number.

Makes navigation quicker while playing.